news_embedding/requirements.txt
git+https://github.com/eCAPTION/ecaption_utils.git
pymongo
python-dotenv
